add counting algo

+++ src/main/java/high/skill/girl/project/calories_counter/dto/ProductDto.java
@@ -7,4 +7,4 @@ public record ProductDto(String name,
                          float proteins,
                          float fats,
                          float carbohydrates,
-                         float calories) { }
+                         int calories) { }
